Updated Go installer

+++ bin/installs/go.sh
@@ -16,10 +16,13 @@ version_name="go1.17.4.linux-amd64.tar.gz"
 
     ## Main Script ##
 
-# Download and export Go repository
-wget -O ${download_dir}/go.tar.gz https://go.dev/dl/${version_name}
-sudo rm -rf /usr/local/go && sudo tar -C /usr/local -xzf ${download_dir}/go.tar.gz
-rm ${download_dir}/go.tar.gz
+# Only install if Go doesn't exist
+if [[ ! -e /usr/local/go ]]; then
+    # Download and export Go repository
+    wget -O ${download_dir}/go.tar.gz https://go.dev/dl/${version_name}
+    sudo rm -rf /usr/local/go && sudo tar -C /usr/local -xzf ${download_dir}/go.tar.gz
+    rm ${download_dir}/go.tar.gz
+fi
 
 # Install Go binaries
 vim +'silent :GoInstallBinaries' +qall
